ABC-319: Script is renamed to index.js. File is documented
