+++ package.json
@@ -1,26 +1,16 @@
 {
   "name": "prefix-commit-jira-id",
   "version": "1.0.0",
-  "description": "",
+  "description": "Adds jira id prefix for each git commit",
   "main": "index.js",
   "scripts": {
-    "commitmsg": "node ./scripts/commit-msg.js",
+    "commitmsg": "node ./index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "cz-jira-smart-commit": "^3.0.0",
-    "lodash": "^4.17.10"
-  },
-  "devDependencies": {
-    "@commitlint/cli": "^6.1.3",
-    "@commitlint/config-conventional": "^6.1.3",
-    "commitizen": "^2.9.6",
     "husky": "^0.14.3",
-    "simple-commit-message": "^4.0.3"
-  },
-  "release": {
-    "analyzeCommits": "simple-commit-message"
+    "lodash": "^4.17.10"
   }
 }
